Further improved UI, and implemented keyboard input.

+++ Include/Utils/DataConversions.hpp
@@ -4,7 +4,7 @@
 
 namespace SHG
 {
-	std::string ConvertToHexString(uint16_t data, int width);
+	std::string ConvertToHexString(uint16_t data, uint8_t width);
 	std::string GetHexString8(uint8_t data);
-	std::string GetHexString16(uint8_t data);
+	std::string GetHexString16(uint16_t data);
 }
